commit avance

+++ src/main/java/com/digis01/ISanchezPruebaCajero/DAO/CuentaRepository.java
@@ -1,6 +1,7 @@
 package com.digis01.ISanchezPruebaCajero.DAO;
 
 import com.digis01.ISanchezPruebaCajero.JPA.Cuenta;
+import java.math.BigDecimal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.query.Procedure;
 import org.springframework.data.repository.query.Param;
@@ -10,5 +11,5 @@ import org.springframework.stereotype.Repository;
 public interface CuentaRepository extends JpaRepository<Cuenta, Integer>{
     
     @Procedure(procedureName = "GetMontoUsuario", outputParameterName = "pMontoCuenta")
-    Long getMontoCuentaUsuario(@Param("pIdUsuario") int idUsuario);
+    BigDecimal getMontoCuentaUsuario(@Param("pIdUsuario") int idUsuario);
 }
